Improve performance by 40.98% by improving `System.read`
Replaced some of the many if-conditions in `System.read` with a switch statement on the last 4 bits of the address.

This improved performance by 40.98%, as the switch is not only faster but is also executed way earlier.

The testing was done by running the first 25 million M-cycles of the `cpu_instrs` blaarg test, which so happens to be around when the test actually finishes. The measure was taken using `window.performance.now()`, before and after the frame was drawn, and summing the values. This ensures that all instructions are run, and that they are run a sufficient amount of times.

The results were:
Time taken before the change: 33 955.9ms
Time taken after the change: 20 039.1ms
Difference = (after - before)/before = -0.4098

Further improvements can be done by adding a second switch-statements on the following 4 bytes of the address, and repeating the process. However for now this is more than enough of an improvement.